4vol, a single-speaker mono audio power quadrupler for JACK
I only have one working speaker (and a very nice one too!) but driven
from the headphone output of my laptop it is rather quiet and, of course,
you only hear one channel. 4vol averages the two channels to make mono and
feeds plus this to the left channel and minus the same to the right.
The speaker is connected between the left and right signal leads from
the headphone output and gets twice the voltage cwbetween its terminals
hence four times the power.